pyproject.toml
[tool.poetry]
name = "battleship"
version = "0.1"
description = ""
authors = ["serg <sergwle@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.6"
requests = "^2.28.1"
numpy = "^1.23.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
django-rest-framework = "^0.1.0"
Django = "^4.1"
"backports.weakref" = "^1.0.post1"
"backports.zoneinfo" = "^0.2.1"
flake8 = "^5.0.4"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
